requirements.txt
Pillow>=10.0.0
pyautogui>=0.9.54
numpy>=1.24.0
keyboard>=0.13.5
opencv-python>=4.8.0
scikit-learn>=1.3.0